requirements.txt
pandas
datasets
pytest
groq
python-dotenv
fastapi
uvicorn
httpx
streamlit
